Rewrite heal location data reading

+++ include/core/heallocation.h
@@ -11,17 +11,17 @@ class HealLocation {
 
 public:
     HealLocation()=default;
-    HealLocation(QString, QString, int, uint16_t, uint16_t, QString = "", uint16_t = 0);
+    HealLocation(QString, QString, int, int16_t, int16_t, QString = "", uint8_t = 0);
     friend QDebug operator<<(QDebug debug, const HealLocation &hl);
 
 public:
     QString idName;
     QString mapName;
     int     index;
-    uint16_t  x;
-    uint16_t  y;
+    int16_t  x;
+    int16_t  y;
     QString respawnMap;
-    uint16_t respawnNPC;
+    uint8_t respawnNPC;
     static HealLocation fromEvent(Event *);
 };
