+++ src/core/heallocation.cpp
@@ -4,8 +4,8 @@
 #include "map.h"
 
 HealLocation::HealLocation(QString id, QString map,
-                           int i, uint16_t x, uint16_t y,
-                           QString respawnMap, uint16_t respawnNPC) {
+                           int i, int16_t x, int16_t y,
+                           QString respawnMap, uint8_t respawnNPC) {
     this->idName = id;
     this->mapName = map;
     this->index = i;
